+++ src/main/kotlin/com/easternkite/remain/Application.kt
@@ -1,6 +1,7 @@
 package com.easternkite.remain
 
 import com.easternkite.remain.features.fortune.fortuneRoute
+import com.easternkite.remain.features.time.configureRouting
 import io.ktor.server.application.*
 import io.ktor.server.engine.*
 import io.ktor.server.netty.*
