--- src/main/kotlin/com/easternkite/remain/enum/Keyword.kt
@@ -1,5 +1,0 @@
-package com.easternkite.remain.enum
-
-enum class Keyword(val value: String) {
-    CLOCK_OUT("퇴근"),
-}
