+++ src/main/kotlin/com/easternkite/remain/features/time/TimeKeyword.kt
@@ -1,0 +1,5 @@
+package com.easternkite.remain.features.time
+
+enum class TimeKeyword(val value: String) {
+    CLOCK_OUT("퇴근"),
+}
