+++ src/main/kotlin/com/easternkite/remain/model/DrResponse.kt
@@ -1,0 +1,9 @@
+package com.easternkite.remain.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DrResponse(
+    val text: String,
+    val responseType: String
+)
